trabalho03/gateway/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
customtkinter==5.2.1
darkdetect==0.8.0
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httptools==0.6.0
idna==3.4
packaging==23.2
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pyserial==3.5
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tk==0.1.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
